Stop bundling Robolectric Processor's sdks.txt file as a Java resource
The resources approach would be difficult to get working in Gradle/Maven for
custom shadow packages outside of the Robolectric project itself. This is
because in Gradle/Maven, the the processor jar file currently contains a
sdks.txt resource that has hardcoded paths from the releaser's workstation.

Instead switch to a file-based scheme and emit the sdks.txt file to
Robolectric's root project's build directory. With a file based scheme is is
possible for external projects to specify an sdks.txt file path as well.

Also turn off the ImplementsValidator plugin off by default as it is
non-trivial for custom shadow packages in the Gradle/Maven world to generate
sdks.txt files. This validation is primarily intended for the Robolectric
project itself.

Fixes #4801

PiperOrigin-RevId: 344144378